% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cm_to_inch}
\alias{cm_to_inch}
\title{Convert Centimeter to Inches}
\usage{
cm_to_inch(x, unclass = TRUE)
}
\arguments{
\item{x}{A numeric vector as Centimeter}

\item{unclass}{If \code{FALSE} return object class "units"}
}
\value{
A numeric vector as Inches
}
\description{
Convert Centimeter to Inches
}
\examples{
cm_to_inch(1:3)
}
