% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot_df.R
\name{cor.mtest}
\alias{cor.mtest}
\title{Helper function for corrplot_df}
\usage{
cor.mtest(mat, method)
}
\arguments{
\item{mat}{data frame only numeric column}

\item{method}{method for calc p-value}
}
\value{
matrix of p-value
}
\description{
Helper function for corrplot_df
}
