% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot_df.R
\name{corrplot_df}
\alias{corrplot_df}
\title{Correlation plot from a data frame}
\usage{
corrplot_df(
  df,
  ...,
  method_cor = "pearson",
  method_plot = "custom",
  na.rm_type = "rows",
  type = "upper",
  tl.srt = 45,
  tl.col = "black",
  sig.level = 0.05,
  insig = "blank"
)
}
\arguments{
\item{df}{Data frame}

\item{...}{Pass to \code{corrplot::corrplot()} argument}

\item{method_cor}{Character: Specify method for calculate
correlation; "pearson" (default), "spearman" or "kendall"}

\item{method_plot}{Character: Specify plotting output;
"custom" (default) other options are "circle", "square", "ellipse", "number", "shade",
"color", "pie" .}

\item{na.rm_type}{Character: If NA's is present in data frame,
it will be removed by "rows"(default): any rows that has NA's or "cols" any column that has NA's.}

\item{type}{Character: "upper" (default), "lower" or "full", display full matrix, lower triangular or upper triangular matrix.}

\item{tl.srt}{Numeric: For text label string rotation in degrees.}

\item{tl.col}{Character: Specify text color "black" (default)}

\item{sig.level}{Numeric: P-value significant level, default = 0.05.}

\item{insig}{Character: Character, specialized insignificant correlation coefficients,
"pch" (default), "p-value", "blank", "n", or "label_sig". If "blank", wipe away the corresponding glyphs;
if "p-value", add p-values the corresponding glyphs; if "pch", add characters (see pch for details) on corresponding glyphs;
if "n", don't take any measures; if "label_sig", mark significant correlations with pch (see sig.level).}
}
\value{
Plot
}
\description{
Correlation plot from a data frame, a wrapper around \code{corrplot::corrplot()}
}
\examples{
# Plot correlation of numeric column
corrplot_df(iris)

mtcars \%>\%
  corrplot_df(method_plot = "circle", # specify graphic method of correlation
             type = "lower",         # display lower half of correlation plot
             sig.level = 0.01,       # p-value = 0.01
             insig = "pch")          # insignificant p-value will be crossed out


}
\seealso{
\code{corrplot} package
\url{https://cran.r-project.org/web/packages/corrplot/vignettes/corrplot-intro.html}
}
