% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{count_vars}
\alias{count_vars}
\title{Count Variables in a data.frame}
\usage{
count_vars(x, ..., sort = TRUE, name = NULL)
}
\arguments{
\item{x}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{...}{Variables to count (one level); If not supply, count all variables.}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}

\item{name}{The name of the new column in the output.}
}
\value{
A named list of data.frame
}
\description{
Count all or many variables in a data.frame and display the results as list of data.frame.
}
\examples{
# Count All variables
count_vars(iris)
# Count Specified variables
count_vars(iris, Species, Sepal.Length)
}
