% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fig.mac}
\alias{fig.mac}
\title{figure height & weight config for Macbook pro}
\format{
A data frame
\describe{
\item{wd.cm}{width of figure (cm.)}
\item{ht.cm}{height of figure (cm.)}
\item{wd}{width of figure (inch)}
\item{ht}{height of figure (inch)}
\item{cm.to.inch}{Conversion factor from cm to inch}
}
\usage{
fig.mac
}
\description{
figure height & weight config for Macbook pro
}
\keyword{datasets}
