% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{geom_arrow_pca}
\alias{geom_arrow_pca}
\title{Arrow for PCA bi-plot}
\usage{
geom_arrow_pca(
  pca_recipe,
  x = PC1,
  y = PC2,
  x_origin = 0,
  y_origin = 0,
  length = ggplot2::unit(0.05, "inches"),
  type = "closed",
  color = "black",
  ...
)
}
\arguments{
\item{pca_recipe}{Object class recipe that already \code{step_pca}}

\item{x}{unquoted name to specify which principle component to plot in x-axis (must be one of PC1, PC2, PC3, ... ,etc.)}

\item{y}{unquoted name to specify which principle component to plot in y-axis (must be one of PC1, PC2, PC3, ... ,etc.)}

\item{x_origin}{(numeric) x-origin of the arrow}

\item{y_origin}{(numeric) y-origin of the arrow}

\item{length}{length of the arrow defined by function \code{ggplot2::unit} (passed to \code{ggplot2::arrow})}

\item{type}{type of the arrow head (passed to \code{ggplot2::arrow})}

\item{color}{color of the arrow segment (passed to \code{ggplot2::geom_segment})}

\item{...}{passed to \code{geom_segment}}
}
\value{
plot arrow
}
\description{
Arrow for PCA bi-plot
}
\examples{
library(magrittr)
library(lbr)
prep_pca(iris) \%>\%
  pca_scatter_plot(ggplot2::aes(color = Species)) +
  geom_arrow_pca(prep_pca(iris))
}
