% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{geom_label_pca}
\alias{geom_label_pca}
\title{Vector's Label for PCA bi-plot}
\usage{
geom_label_pca(
  pca_recipe,
  x = PC1,
  y = PC2,
  geom_type = geom_label,
  hjust = 0,
  vjust = 1,
  size = 3,
  color = "#0A537D",
  check_overlap = FALSE,
  label.padding = ggplot2::unit(0.25, "lines"),
  label.r = ggplot2::unit(0.15, "lines"),
  label.size = 0.25,
  segment.color = "black",
  min.segment.length = 0.5,
  force = 1,
  force_pull = 1,
  max.overlaps = 10,
  size_ggrepel = 2,
  ...
)
}
\arguments{
\item{pca_recipe}{Object class recipe that already \code{step_pca}}

\item{x}{unquoted name to specify which principle component to plot in x-axis (must be one of PC1, PC2, PC3, ... ,etc.)}

\item{y}{unquoted name to specify which principle component to plot in y-axis (must be one of PC1, PC2, PC3, ... ,etc.)}

\item{geom_type}{unquoted name of geom to use, must be one of \code{geom_label}, \code{geom_label_repel}, \code{geom_text}, or \code{geom_text_repel}}

\item{hjust}{(numeric) passed to \code{geom_text} or \code{geom_label}}

\item{vjust}{(numeric) passed to \code{geom_text} or \code{geom_label}}

\item{size}{(numeric) passed to \code{geom_text} or \code{geom_label}}

\item{color}{passed to \code{geom_text} or \code{geom_label}}

\item{check_overlap}{passed to \code{geom_text}}

\item{label.padding}{passed to \code{geom_label}}

\item{label.r}{passed to \code{geom_label}}

\item{label.size}{passed to \code{geom_label}}

\item{segment.color}{passed to \code{geom_label_repel} or \code{geom_text_repel}}

\item{min.segment.length}{passed to \code{geom_label_repel} or \code{geom_text_repel}}

\item{force}{passed to \code{geom_label_repel} or \code{geom_text_repel}}

\item{force_pull}{passed to \code{geom_label_repel} or \code{geom_text_repel}}

\item{max.overlaps}{passed to \code{geom_label_repel} or \code{geom_text_repel}}

\item{size_ggrepel}{passed to \code{geom_label_repel} or \code{geom_text_repel}}

\item{...}{passed to geom that specified by \code{geom_type} argument}
}
\value{
plot label
}
\description{
Vector's Label for PCA bi-plot
}
\examples{
library(magrittr)
library(lbr)
prep_pca(iris) \%>\%
  pca_scatter_plot(ggplot2::aes(color = Species)) +
  geom_label_pca(prep_pca(iris))
}
