% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggsave_mac}
\alias{ggsave_mac}
\title{My custom ggsave for macbook pro}
\usage{
ggsave_mac(
  filename,
  plot = last_plot(),
  path = NULL,
  width = 25.6,
  height = 16,
  units = "cm",
  dpi = "retina",
  scale = 1,
  device = NULL,
  ...
)
}
\arguments{
\item{filename}{(charater)     File name to create on disk.}

\item{plot}{\code{last_plot()} passed to ggsave}

\item{path}{Path of the directory to save plot to: path and filename are combined to create the fully qualified file name. Defaults to the working directory.}

\item{width}{Plot size in units ("in", "cm", or "mm").}

\item{height}{Plot size in units ("in", "cm", or "mm").}

\item{units}{Plot size in units ("in", "cm", or "mm").}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320), "print" (300), or "screen" (72). Applies only to raster output types.}

\item{scale}{Multiplicative scaling factor.}

\item{device}{Device to use. Can either be a device function (e.g. png()), or one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{...}{passed to \code{ggsave}}
}
\description{
My custom ggsave for macbook pro
}
\examples{
# ggsave_mac("path/to/plot.png")
}
