% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google.R
\name{gs4_create_at}
\alias{gs4_create_at}
\title{Create Google Sheet at Specified Location}
\usage{
gs4_create_at(
  path = NULL,
  name = gs4_random(),
  sheets = NULL,
  overwrite = NA,
  ...
)
}
\arguments{
\item{path}{Specifies target destination for the file on Google Drive. Can be an actual path (character), a file id marked with as_id(), or a dribble.}

\item{name}{The name of the new spreadsheet.}

\item{sheets}{Optional input for initializing (work)sheets. If unspecified, the Sheets API automatically creates an empty "Sheet1". You can provide a vector of sheet names, a data frame, or a (possibly named) list of data frames.}

\item{overwrite}{\itemize{
\item \code{NA} (default): Just do the operation, even if it results in multiple files with the same filepath.
\itemize{
\item \code{TRUE}: Check for a pre-existing file at the filepath. If there is zero or one, move a pre-existing file to the trash, then carry on. Note that the new file does not inherit any properties from the old one, such as sharing or publishing settings. It will have a new file ID. An error is thrown if two or more pre-existing files are found.
\item \code{FALSE}: Error if there is any pre-existing file at the filepath.
}
}}

\item{...}{to \code{gs4_create()}: Optional spreadsheet properties that can be set through this API endpoint, such as locale and time zone.}
}
\value{
Object class: "sheets_id", "drive_id"
}
\description{
Calls \code{gs4_create()} then move file to specified location with \code{drive_mv()}
}
\examples{
# gs4_create_at("folder_name", name = "file_name", sheets = "sheet_name")
}
