% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inch_to_cm}
\alias{inch_to_cm}
\title{Convert Inches to Centimeter}
\usage{
inch_to_cm(x, unclass = TRUE)
}
\arguments{
\item{x}{A numeric vector as Inches}

\item{unclass}{If \code{FALSE} return object class "units"}
}
\value{
A numeric vector as Centimeter
}
\description{
Convert Inches to Centimeter
}
\examples{
inch_to_cm(2.5)
}
