% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBR-main.R
\docType{package}
\name{LBR-package}
\alias{LBR-package}
\alias{leerLBRDatos}
\title{Lee las bases de datos de biología reproductiva}
\usage{
leerLBRDatos(filename, useAll = TRUE, initialDates = NULL,
  finalDates = NULL, listMonths = NULL, stringsAsFactors = FALSE, ...)
}
\arguments{
\item{filename}{Archivo conteniendo la base de datos a trabajar.}

\item{...}{Argumentos extra, exportados a la función \code{read.csv}.}
}
\description{
Funcion para leer bases de datos de biología reproductiva desde un directorio
especificado.
}
\details{
LBR

\tabular{ll}{ Package: \tab LBR\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2016-05-01\cr License: \tab GPL-2\cr }
}
\examples{

}

