% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBR-main.R
\name{obtenerIndices}
\alias{obtenerIndices}
\title{Obtiene un objeto con los principales índices reproductivos: IGS, FC, FD}
\usage{
obtenerIndices(data, sp = "anc", ...)
}
\arguments{
\item{data}{Objeto de clase 'LBRData' conteniendo información para el cálculo de
IGS (peso gónada, peso eviscerado), FC (pero eviscerado, longitud total)}

\item{...}{Argumentos extra, exportados a la función \code{read.csv}.}
}
\description{
Funcion para obtener un objeto de clase LBRindex con los principales índices para
una especie indicada: Índice Gonadosomático, Factor de Condición, Fracción Desovante
}
\examples{
obtenerIndices(filename)
}

