% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{parse_to_formula}
\alias{parse_to_formula}
\title{Parse symbols to LHS and RHS of formula}
\usage{
parse_to_formula(fun, lhs, rhs, ...)
}
\arguments{
\item{fun}{Function that accept formula object as first argument}

\item{lhs}{(Quote or Unquoted) Expression for left-hand sided of the formula}

\item{rhs}{(Quote or Unquoted) Expression for right-hand sided of the formula}

\item{...}{pass to \code{fun}}
}
\value{
object as return by \code{fun}
}
\description{
Most common use cases would be passing string to \code{lhr} and \code{rhs} of the formula.
}
\examples{
parse_to_formula(lm, "Sepal.Length", "Sepal.Width", data = iris)
}
