% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_biplot}
\alias{pca_biplot}
\title{PCA Bi-plot}
\usage{
pca_biplot(
  pca_recipe,
  mapping = NULL,
  x = PC1,
  y = PC2,
  alpha = 0.8,
  size = 2,
  geom_type_scatter = geom_point,
  x_origin = 0,
  y_origin = 0,
  arrow_scale = 1,
  color_arrow = "black",
  geom_type = geom_label,
  hjust = 0,
  vjust = 1,
  size_txt = 3,
  color = "#0A537D",
  check_overlap = FALSE,
  label.padding = ggplot2::unit(0.25, "lines"),
  label.r = ggplot2::unit(0.15, "lines"),
  label.size = 0.25,
  segment.color = "black",
  min.segment.length = 0.5,
  force = 1,
  force_pull = 1,
  max.overlaps = 10,
  size_ggrepel = 2,
  ...
)
}
\arguments{
\item{pca_recipe}{Object class recipe that already \code{step_pca}}

\item{mapping}{\code{ggplot2} aesthetic mapping (passed to \code{geom_point})}

\item{x}{unquoted name to specify which principle component to plot in x-axis (must be one of PC1, PC2, PC3, ... ,etc.)}

\item{y}{unquoted name to specify which principle component to plot in y-axis (must be one of PC1, PC2, PC3, ... ,etc.)}

\item{alpha}{point transparency (passed to \code{geom_point})}

\item{size}{size of point (passed to \code{geom_point})}

\item{geom_type_scatter}{(unquoted name) Type of scatter plot: \code{geom_point} or \code{geom_text}}

\item{x_origin}{(numeric) x-origin of the arrow (passed to \code{lbr::geom_arrow_pca})}

\item{y_origin}{(numeric) y-origin of the arrow (passed to \code{lbr::geom_arrow_pca})}

\item{arrow_scale}{(numeric) scale factor of the arrow (passed to \code{lbr::geom_arrow_pca})}

\item{color_arrow}{(character) color of the arrow (passed to \code{lbr::geom_arrow_pca})}

\item{geom_type}{unquoted name of geom to use, must be one of \code{geom_label}, \code{geom_label_repel}, \code{geom_text}, or \code{geom_text_repel}}

\item{hjust}{(numeric) passed to \code{geom_text} or \code{geom_label}}

\item{vjust}{(numeric) passed to \code{geom_text} or \code{geom_label}}

\item{size_txt}{(numeric) size of \code{geom_text} or \code{geom_label}}

\item{color}{(character) color of \code{geom_text} or \code{geom_label}}

\item{check_overlap}{passed to \code{geom_text}}

\item{label.padding}{passed to \code{geom_label}}

\item{label.r}{passed to \code{geom_label}}

\item{label.size}{passed to \code{geom_label}}

\item{segment.color}{color of the segment, \code{NA} for no line segment (passed to \code{geom_label_repel} or \code{geom_text_repel})}

\item{min.segment.length}{minimum length that will draw line segment (passed to \code{geom_label_repel} or \code{geom_text_repel})}

\item{force}{pushing force for overlapping label (passed to \code{geom_label_repel} or \code{geom_text_repel})}

\item{force_pull}{pulling force to data points (passed to \code{geom_label_repel} or \code{geom_text_repel})}

\item{max.overlaps}{maximum label overlapping (passed to \code{geom_label_repel} or \code{geom_text_repel})}

\item{size_ggrepel}{size of the label (passed to \code{geom_label_repel} or \code{geom_text_repel})}

\item{...}{passed to \code{geom_point}}
}
\value{
plot
}
\description{
Plot PCA bi-plot from \code{recipe} object that has \code{step_pca}
}
