% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_scatter_plot}
\alias{pca_scatter_plot}
\title{PCA Scatter Plot}
\usage{
pca_scatter_plot(
  pca_recipe,
  mapping = NULL,
  x = PC1,
  y = PC2,
  geom_type = geom_point,
  alpha = 0.8,
  size = 2,
  ...
)
}
\arguments{
\item{pca_recipe}{object class recipe that already \code{step_pca}}

\item{mapping}{aesthetic i.e. \code{ggplot2::aes()} mapping}

\item{x}{unquoted name to specify which principle component to plot in x-axis (must be one of PC1, PC2, PC3, ... ,etc.)}

\item{y}{unquoted name to specify which principle component to plot in y-axis (must be one of PC1, PC2, PC3, ... ,etc.)}

\item{geom_type}{unquoted name to specify type of geom to plot: \code{geom_point} or \code{geom_text}}

\item{alpha}{passed to \code{geom_point}}

\item{size}{passed to \code{geom_point}}

\item{...}{passed to \code{geom_point}}
}
\value{
plot
}
\description{
PCA Scatter Plot
}
\examples{
library(magrittr)
library(lbr)
iris \%>\%
  prep_pca() \%>\%
  pca_scatter_plot(ggplot2::aes(color = Species))
}
