% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_scree_plot}
\alias{pca_scree_plot}
\title{PCA Screeplot}
\usage{
pca_scree_plot(
  pca_recipe,
  terms = "percent_var",
  title = "Scree Plot",
  y = NA,
  ...
)
}
\arguments{
\item{pca_recipe}{Object class recipe that already \code{step_pca}}

\item{terms}{(character) To specify the y-axis, must be one of "var" for variance, "percent_var" for percent variance, and "cum_var" for cumulative percent variance}

\item{title}{(character) Title of the plot}

\item{y}{Y-axis label: \code{NA} for default label, character to specify your own label, or \code{NULL} to remove label}

\item{...}{passed to \code{ggplot2::labs}}
}
\value{
plot
}
\description{
PCA Screeplot
}
\examples{
library(magrittr)
library(lbr)
iris \%>\%
  prep_pca() \%>\%
  pca_scree_plot()
}
