% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{prep_pca}
\alias{prep_pca}
\title{Automatic Recipes for PCA}
\usage{
prep_pca(df, fct_vars = NULL, num_vars = NULL)
}
\arguments{
\item{df}{data frame}

\item{fct_vars}{(character) Factor variables to choose as ID (if \code{NULL}, choose all factor variables in data frame)}

\item{num_vars}{(character) Numeric variables to choose as feature (if \code{NULL}, choose all numeric variables in data frame)}
}
\value{
recipe object
}
\description{
use \code{recipes} package to \code{step_nzv},\code{step_naomit}, \code{step_normalize}, \code{step_pca}, and \code{prep}
}
\examples{
prep_pca(iris)
}
