% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R, R/read.dir.R
\name{print_messages}
\alias{print_messages}
\title{Helper function}
\usage{
print_messages(chr_vec)

print_messages(chr_vec)
}
\arguments{
\item{chr_vec}{character}
}
\value{
messages

message
}
\description{
Helper function

Message function
}
