% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dir_csv.R
\name{print_warnings}
\alias{print_warnings}
\title{Helper function : for read_dir_csv}
\usage{
print_warnings(chr_vec)
}
\arguments{
\item{chr_vec}{character}
}
\value{
warning
}
\description{
Helper function : for read_dir_csv
}
