% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{set_names_match}
\alias{set_names_match}
\title{Set Names in an Object that Match Lookup table}
\usage{
set_names_match(x, lookup = x, invert = FALSE, quiet = FALSE, ...)
}
\arguments{
\item{x}{Vector to name.}

\item{lookup}{Lookup vector: the names of \code{lookup} must corresponding to "old names" of \code{x} and the
values of \code{lookup} will be set as "new names" of \code{x}}

\item{invert}{Logical: If \code{TRUE}, invert names and values of \code{lookup}.}

\item{quiet}{Logical: If \code{FALSE}, the message of non-matching names of \code{x} will be displayed.}

\item{...}{passed to \code{rlang:set_names()}}
}
\value{
\code{x} with new names
}
\description{
Set names of object by matching object's old names to names of lookup table.
This will ensure that object's names will always named correctly.
This function also wraps \code{rlang::set_names()}.
}
\examples{
set_names_match(iris, c("Sepal.Length" = "SL",
                        "Sepal.Width" = "SW",
                        "Petal.Length" = "PL")
               )
}
