% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sourceDir}
\alias{sourceDir}
\title{Source multiple R script from a directory}
\usage{
sourceDir(path, local = knitr::knit_global(), trace = TRUE, ...)
}
\arguments{
\item{path}{(character) Specify directory where .R script located}

\item{local}{passed to \code{source}}

\item{trace}{(logical) whether to display file names message}

\item{...}{passed to \code{source}}
}
\value{
message (or none)
}
\description{
Source multiple R script from a directory
}
\examples{
# sourceDir("path/to/dir")
}
