% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usethis.R
\name{wrap_quote}
\alias{wrap_quote}
\title{Wrap Quote to Text}
\usage{
wrap_quote(.chr)
}
\arguments{
\item{.chr}{Character vector containing texts to wrap}
}
\value{
A message to copy
}
\description{
Useful for turning texts to quoted variable names. The output will ignore any white spaces,
new lines, or any punctuation (except underscore and dot).
}
\examples{
wrap_quote("var_1, !var.2,  ,  `var3`")
}
