% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbr-package.R
\docType{package}
\name{lbr-package}
\alias{lbr}
\alias{lbr-package}
\title{lbr: Data wrangling and analysis}
\description{
My personal taste of data wrangling and analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Lightbridge-AI/lbr}
}

}
\author{
\strong{Maintainer}: Kittipos Sirivongrungson \email{ki11ip0.s.a.s@gmail.com}

}
\keyword{internal}
